+++ src/main/java/br/github/pedroguimaraes/dao/Driverdao.java
@@ -12,12 +12,12 @@ public class Driverdao extends DAO<Driver> {
     }
 
     @Override
-    public Driver getById(Long id) {
+    public Driver getById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public boolean removeById(Long id) {
+    public boolean removeById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
