+++ src/main/java/br/github/pedroguimaraes/dao/Examdao.java
@@ -10,12 +10,12 @@ public class Examdao extends DAO{
     }
 
     @Override
-    public Object getById(Long id) {
+    public Object getById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public boolean removeById(Long id) {
+    public boolean removeById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
