+++ src/main/java/br/github/pedroguimaraes/dao/Vehicledao.java
@@ -12,12 +12,12 @@ public class Vehicledao extends DAO<Vehicle>{
     }
 
     @Override
-    public Vehicle getById(Long id) {
+    public Vehicle getById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public boolean removeById(Long id) {
+    public boolean removeById(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
